requirements.txt
<<<<<<< HEAD
fastapi==0.103.2
uvicorn==0.23.2
arxiv==1.4.3
sentence-transformers==2.2.2
pandas==2.0.3
aiohttp==3.8.6
lxml==4.9.3
biopython==1.81
nest-asyncio==1.5.8
scikit-learn==1.3.0
numpy==1.26.0
gunicorn==21.2.0
=======
fastapi>=0.85.0
uvicorn>=0.19.0
arxiv>=1.4.3
sentence-transformers>=2.2.2
pandas>=1.5.0
aiohttp>=3.8.3
lxml>=4.9.2
biopython>=1.80
nest-asyncio>=1.5.6
scikit-learn>=1.2.0
numpy>=1.24.0
python-dotenv>=0.21.0
>>>>>>> d3b2a1d15dda74c6cfcd04dc1bbcf3d7d5b2924a
